backend/requirements.txt
fastapi
uvicorn[standard]
ultralytics
easyocr
opencv-python
pillow
python-multipart
numpy
